Moves logic and functions of password update form to new child component. New component renders succesfully.
